+++ src/scripts/components/API/forecastAPI/forecastApi.js
@@ -46,7 +46,7 @@ export function getForecasts(latitude, longitude) {
     },
   ];
 
-  const responses = mockUrls.map((url) => {
+  const responses = urls.map((url) => {
     url.link = fetch(url.link, {
       headers: url.headers,
     }).then((res) => res.json());
